move track search hooks into content/search
