+++ src/app/hooks/index.ts
@@ -1,7 +1,5 @@
 import trackPageView from './trackPageView';
-import trackSearch from './trackSearch';
 
 export {
   trackPageView,
-  trackSearch,
 };
